pyproject.toml
[project]
name = "memory-layer"
version = "0.1.0"
description = "Persistent memory service for local agent workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "chromadb>=0.4.0",
  "mcp>=1.0.0",
  "pydantic>=2.0.0",
  "pyyaml>=6.0.0",
  "sentence-transformers>=2.6.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.2.0",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"
